requirements.txt
fastapi
uvicorn
httpx
python-dotenv
aiocache
qdrant-client
numpy
pydantic
nest_asyncio
celery 
redis